Split GUI waveform playback tests
